test(ui): isolate replace-mode catalog e2e coverage
